Don't import tzlocal at top level
tzlocal 5 now uses logging, which means that rfc5424 gets a recursive import error with tzlocal 5.0.1 and later.
This PR solves that, and also does not call tzlocal every time a message is formatted. tzlocals get_localzone()
does cache the result, so it's not a big burden, but it's still not necessary.